+++ netbox_ssl/models/__init__.py
@@ -1,9 +1,15 @@
 from .assignments import CertificateAssignment
-from .certificates import Certificate, CertificateAlgorithmChoices, CertificateStatusChoices
+from .certificates import (
+    Certificate,
+    CertificateAlgorithmChoices,
+    CertificateStatusChoices,
+    ChainStatusChoices,
+)
 
 __all__ = [
     "Certificate",
     "CertificateStatusChoices",
     "CertificateAlgorithmChoices",
+    "ChainStatusChoices",
     "CertificateAssignment",
 ]